Fix menu breaking when Roo is moved between primary and secondary sidebars (#4045)
* Fix menu breaking when Roo is moved between primary and secondary sidebars

Hello Roo Team! We changed this on the Kilo side and thought it might be useful to you!

The menu buttons (Settings etc.) stop working when Roo is moved between the primary and secondary sidebars.
This is because ClineProvider is prematurely disposed in that case
This change prevents the ClineProvider from being disposed when hosted in a sidebar.
It should still be disposed when hosted in a tab, because they have their own ClineProvider instance.

Found while investigating https://github.com/Kilo-Org/kilocode/issues/502.

* refactor: improve logging

* fix: extra bracket

---------

Co-authored-by: Daniel <57051444+daniel-lxs@users.noreply.github.com>